Add UI to site
This commit adds buttons for user to select their choice instead of a
prompt. Round and game results are also displayed on the webpage along
with a round message. Game winning message has increased size, bold font
and a different color depending on if the user or the computer won.